+++ NEWS.md
@@ -8,8 +8,8 @@
 
 * Package pre-release 
 
-    * __v 0.0.14__ [2025-10-04] added fitTubeModel; added deseasonTubeData; 
-    added extractTubeMeta; importing mgcv. 
+    * __v 0.0.14__ [2025-10-10] added fitTubeModel, deseasonTubeData, 
+    extractTubeMeta, checkTubeMeta, repairTubeMeta;importing mgcv. 
     * __v 0.0.13__ [2025-09-01] added updated version of tubeMap.
     * __v 0.0.12__ [2025-08-29] (requested change) tagTubeDates renamed 
     tagTubeStartEnd; added updated version of tubePlot.